requirements.txt
pytest>=8.0.0
playwright>=1.45.0
pytest-playwright>=0.5.0
python-dotenv>=1.0.1
